Refactor : 경로 설정
