+++ src/components/auth/LoginForm.tsx
@@ -1,27 +1,27 @@
 import { changeField, initializeForm } from '@module/auth';
-import { login, register } from '@api/auth';
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import AuthForm from '../../components/auth/AuthForm';
-import { rootState } from '@module';
+import AuthForm from './AuthForm';
+import { login } from '@api/auth';
+import { rootState } from '@module/root';
 import { tempSetUser } from '@module/user';
 import { useAlert } from 'react-alert';
 import { useHistory } from 'react-router-dom';
 
-const RegisterForm = () => {
-  const alert = useAlert();
+const LoginForm = () => {
   const history = useHistory();
+  const alert = useAlert();
   const dispatch = useDispatch();
   const { form } = useSelector(({ auth }: rootState) => ({
-    form: auth.register,
+    form: auth.login,
   }));
   // 인풋 변경 이벤트 핸들러
   const onChange = (e) => {
     const { value, name } = e.target;
     dispatch(
       changeField({
-        form: 'register',
+        form: 'login',
         key: name,
         value,
       })
@@ -31,33 +31,29 @@ const RegisterForm = () => {
   // 폼 등록 이벤트 핸들러
   const onSubmit = async (e) => {
     e.preventDefault();
-    const result = {
-      nickName: form.nickName,
-      userId: form.userId,
-      password: form.password,
-    };
+    if (!form) return;
     try {
-      const response = await register(result);
       const { nickName, status, userId } = await login({
-        userId: response.userId,
+        userId: form.userId,
         password: form.password,
       });
-      dispatch(tempSetUser({ user: { nickName, status, userId } }));
-      history.push('/');
+      dispatch(tempSetUser({ nickName, status, userId }));
+      if (status === 'admin') history.push('/admin');
+      else history.push('/');
     } catch {
-      alert.error('회원가입 실패');
+      alert.error('로그인 실패');
     }
   };
 
   // 컴포넌트가 처음 렌더링될 때 form을 초기화함
   useEffect(() => {
-    dispatch(initializeForm('register'));
-    return () => dispatch(initializeForm('register'));
+    dispatch(initializeForm('login'));
+    return () => dispatch(initializeForm('login'));
   }, [dispatch]);
 
   return (
     <AuthForm
-      type="register"
+      type="login"
       form={form}
       onChange={onChange}
       onSubmit={onSubmit}
@@ -65,4 +61,4 @@ const RegisterForm = () => {
   );
 };
 
-export default RegisterForm;
+export default LoginForm;
